+++ aspnet-core/src/TestDemo.Application.Shared/WholesaleInputs/Dtos/WholesaleEclDataLoanBookDto.cs
@@ -32,7 +32,7 @@ namespace TestDemo.WholesaleInputs.Dtos
 
 		public double? Month12PD { get; set; }
 
-		public int? DaysPastDue { get; set; }
+		public double? DaysPastDue { get; set; }
 
 		public bool WatchlistIndicator { get; set; }
